+++ cmd/aws.go
@@ -6,6 +6,8 @@ import (
 	"fmt"
 	"strings"
 
+	"github.com/aws/aws-sdk-go/aws/awserr"
+	"github.com/aws/aws-sdk-go/aws/credentials"
 	kitlog "github.com/go-kit/kit/log"
 
 	"github.com/cycloidio/terracognita/aws"
@@ -29,15 +31,28 @@ var (
 		Long:  "Terracognita reads from AWS and generates hcl resources and/or terraform state",
 		PreRun: func(cmd *cobra.Command, args []string) {
 			preRunEOutput(cmd, args)
-			viper.BindPFlag("access-key", cmd.Flags().Lookup("access-key"))
-			viper.BindPFlag("secret-key", cmd.Flags().Lookup("secret-key"))
-			viper.BindPFlag("region", cmd.Flags().Lookup("region"))
+
+			viper.BindPFlag("aws-access-key", cmd.Flags().Lookup("aws-access-key"))
+			viper.BindPFlag("aws-secret-access-key", cmd.Flags().Lookup("aws-secret-access-key"))
+			viper.BindPFlag("aws-default-region", cmd.Flags().Lookup("aws-default-region"))
+
+			viper.BindPFlag("aws-shared-credentials-file", cmd.Flags().Lookup("aws-shared-credentials-file"))
+			viper.BindPFlag("aws-profile", cmd.Flags().Lookup("aws-profile"))
+
 			viper.BindPFlag("tags", cmd.Flags().Lookup("tags"))
+
+			// We define aliases so we have an easier access on the code
+			viper.RegisterAlias("access-key", "aws-access-key")
+			viper.RegisterAlias("secret-key", "aws-secret-key")
+			viper.RegisterAlias("region", "aws-default-region")
 		},
 		PostRunE: postRunEOutput,
 		RunE: func(cmd *cobra.Command, args []string) error {
 			logger := log.Get()
 			logger = kitlog.With(logger, "func", "cmd.aws.RunE")
+
+			loadAWSCredentials()
+
 			// Validate required flags
 			if err := requiredStringFlags("access-key", "secret-key", "region"); err != nil {
 				return err
@@ -98,10 +113,48 @@ func init() {
 	awsCmd.AddCommand(awsResourcesCmd)
 
 	// Required flags
-	awsCmd.Flags().String("access-key", "", "Access Key (required)")
-	awsCmd.Flags().String("secret-key", "", "Secret Key (required)")
-	awsCmd.Flags().String("region", "", "Region to search in, for now * it's not supported (required)")
+	awsCmd.Flags().String("aws-access-key", "", "Access Key (required)")
+	awsCmd.Flags().String("aws-secret-access-key", "", "Secret Key (required)")
+	awsCmd.Flags().String("aws-default-region", "", "Region to search in, for now * is not supported (required)")
+	awsCmd.Flags().String("aws-shared-credentials-file", "", "Path to the AWS credential path")
+	awsCmd.Flags().String("aws-profile", "", "Name of the Profile to use with the Credentials")
 
 	// Filter flags
 	awsCmd.Flags().StringSliceVarP(&tags, "tags", "t", []string{}, "List of tags to filter with format 'NAME:VALUE'")
+}
+
+// loadAWSCredentials will first read from ENV and if AccessKey and SecretAccessKey are not found (both of them)
+// will fallback to the SharedCredentials with the profile
+func loadAWSCredentials() error {
+	creds := credentials.NewCredentials(&credentials.ChainProvider{
+		Providers: []credentials.Provider{
+			&credentials.EnvProvider{},
+			&credentials.SharedCredentialsProvider{Filename: viper.GetString("aws-shared-credentials-file"), Profile: viper.GetString("aws-profile")},
+		},
+	})
+
+	value, err := creds.Get()
+	if err != nil {
+		// The NoCredentialProviders is an error returned by Get to identify that none
+		// of the Providers (credentials.EnvProvider and credentials.SharedCredentialsProvider)
+		// did find any information.
+		// So we escape it means nothing was found by AWS
+		if awsE, ok := err.(awserr.Error); ok && awsE.Code() == "NoCredentialProviders" {
+			return nil
+		}
+		return err
+	}
+
+	// If the values are already set
+	// it'll not be override as they
+	// are more relevant
+	if !viper.IsSet("access-key") {
+		viper.Set("access-key", value.AccessKeyID)
+	}
+
+	if !viper.IsSet("secret-key") {
+		viper.Set("secret-key", value.SecretAccessKey)
+	}
+
+	return nil
 }